+++ radiantcore/brush/Winding.h
@@ -81,7 +81,7 @@ public:
 	// Returns the classification for the given plane
 	BrushSplitType classifyPlane(const Plane3& plane) const;
 
-	static PlaneClassification classifyDistance(const float distance, const float epsilon);
+	static PlaneClassification classifyDistance(const double distance, const double epsilon);
 
 	/// \brief Returns true if
 	/// !flipped && winding is completely BACK or ON
